seo-gen/pyproject.toml
[project]
name = "seo-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "openai>=1.76.0",
    "openai-agents[viz]>=0.0.13",
    "pillow>=11.2.1",
    "requests>=2.32.3",
    "rich>=14.0.0",
]
